Move ifLoggedIn to the hoc folder
